+++ src/Symfony/Component/EventDispatcher/Debug/TraceableEventDispatcherInterface.php
@@ -14,6 +14,8 @@ namespace Symfony\Component\EventDispatcher\Debug;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 
 /**
+ * @deprecated since version 4.1, will be removed in 5.0.
+ *
  * @author Fabien Potencier <fabien@symfony.com>
  */
 interface TraceableEventDispatcherInterface extends EventDispatcherInterface