+++ src/Symfony/Component/HttpKernel/DataCollector/EventDataCollector.php
@@ -11,10 +11,11 @@
 
 namespace Symfony\Component\HttpKernel\DataCollector;
 
+use Symfony\Component\EventDispatcher\Debug\TraceableEventDispatcher;
+use Symfony\Component\EventDispatcher\Debug\TraceableEventDispatcherInterface;
+use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\EventDispatcher\EventDispatcherInterface;
-use Symfony\Component\EventDispatcher\Debug\TraceableEventDispatcherInterface;
 
 /**
  * EventDataCollector.
@@ -38,6 +39,7 @@ class EventDataCollector extends DataCollector implements LateDataCollectorInter
         $this->data = array(
             'called_listeners' => array(),
             'not_called_listeners' => array(),
+            'orphaned_events' => array(),
         );
     }
 
@@ -56,6 +58,11 @@ class EventDataCollector extends DataCollector implements LateDataCollectorInter
             $this->setCalledListeners($this->dispatcher->getCalledListeners());
             $this->setNotCalledListeners($this->dispatcher->getNotCalledListeners());
         }
+
+        if ($this->dispatcher instanceof TraceableEventDispatcher) {
+            $this->setOrphanedEvents($this->dispatcher->getOrphanedEvents());
+        }
+
         $this->data = $this->cloneVar($this->data);
     }
 
@@ -64,7 +71,7 @@ class EventDataCollector extends DataCollector implements LateDataCollectorInter
      *
      * @param array $listeners An array of called listeners
      *
-     * @see TraceableEventDispatcherInterface
+     * @see TraceableEventDispatcher
      */
     public function setCalledListeners(array $listeners)
     {
@@ -76,7 +83,7 @@ class EventDataCollector extends DataCollector implements LateDataCollectorInter
      *
      * @return array An array of called listeners
      *
-     * @see TraceableEventDispatcherInterface
+     * @see TraceableEventDispatcher
      */
     public function getCalledListeners()
     {
@@ -86,9 +93,9 @@ class EventDataCollector extends DataCollector implements LateDataCollectorInter
     /**
      * Sets the not called listeners.
      *
-     * @param array $listeners An array of not called listeners
+     * @param array $listeners
      *
-     * @see TraceableEventDispatcherInterface
+     * @see TraceableEventDispatcher
      */
     public function setNotCalledListeners(array $listeners)
     {
@@ -98,13 +105,37 @@ class EventDataCollector extends DataCollector implements LateDataCollectorInter
     /**
      * Gets the not called listeners.
      *
-     * @return array An array of not called listeners
+     * @return array
      *
-     * @see TraceableEventDispatcherInterface
+     * @see TraceableEventDispatcher
      */
     public function getNotCalledListeners()
     {
         return $this->data['not_called_listeners'];
+    }
+
+    /**
+     * Sets the orphaned events.
+     *
+     * @param array $events An array of orphaned events
+     *
+     * @see TraceableEventDispatcher
+     */
+    public function setOrphanedEvents(array $events)
+    {
+        $this->data['orphaned_events'] = $events;
+    }
+
+    /**
+     * Gets the orphaned events.
+     *
+     * @return array An array of orphaned events
+     *
+     * @see TraceableEventDispatcher
+     */
+    public function getOrphanedEvents()
+    {
+        return $this->data['orphaned_events'];
     }
 
     /**